separate apollo client setup code
